Update to toml++ v3

+++ Sources/TOMLKit/ValueOptions.swift
@@ -5,7 +5,7 @@
 //  The full text of the license can be found in the file named LICENSE.
 
 /// Formatting options for TOML integers.
-public enum ValueOptions: UInt8 {
+public enum ValueOptions: UInt16 {
 	/// The value will not be formatted.
 	case none = 0
